+++ test/NonEmptyArray.ts
@@ -6,7 +6,6 @@ import * as _ from '../src/NonEmptyArray'
 import * as O from '../src/Option'
 import { ordNumber } from '../src/Ord'
 import { semigroupSum } from '../src/Semigroup'
-import { showString } from '../src/Show'
 import * as S from '../src/string'
 
 describe('NonEmptyArray', () => {
@@ -327,9 +326,9 @@ describe('NonEmptyArray', () => {
   })
 
   it('getShow', () => {
-    const S = _.getShow(showString)
-    assert.deepStrictEqual(S.show(['a']), `["a"]`)
-    assert.deepStrictEqual(S.show(['a', 'b', 'c']), `["a", "b", "c"]`)
+    const Sh = _.getShow(S.Show)
+    assert.deepStrictEqual(Sh.show(['a']), `["a"]`)
+    assert.deepStrictEqual(Sh.show(['a', 'b', 'c']), `["a", "b", "c"]`)
   })
 
   it('alt / concat', () => {