+++ test/Set.ts
@@ -1,13 +1,23 @@
 import * as assert from 'assert'
+import { getMonoid } from '../src/Array'
 import { left, right } from '../src/Either'
+import * as Eq from '../src/Eq'
+import { pipe } from '../src/function'
+import { none, some as optionSome } from '../src/Option'
 import { ordNumber } from '../src/Ord'
 import {
   chain,
+  compact,
+  difference,
+  empty,
   every,
   filter,
+  filterMap,
+  foldMap,
   fromArray,
-  getIntersectionSemigroup,
   getEq,
+  getIntersectionSemigroup,
+  getShow,
   getUnionMonoid,
   insert,
   intersection,
@@ -16,25 +26,14 @@ import {
   partitionMap,
   reduce,
   remove,
+  separate,
   singleton,
   some,
   subset,
   toArray,
   toggle,
-  union,
-  difference,
-  compact,
-  separate,
-  filterMap,
-  foldMap,
-  getShow,
-  empty
+  union
 } from '../src/Set'
-import * as Eq from '../src/Eq'
-import { none, some as optionSome } from '../src/Option'
-import { showString } from '../src/Show'
-import { getMonoid } from '../src/Array'
-import { pipe } from '../src/function'
 import * as S from '../src/string'
 
 const gte2 = (n: number) => n >= 2
@@ -279,12 +278,12 @@ describe('Set', () => {
   })
 
   it('getShow', () => {
-    const S = getShow(showString)
+    const Sh = getShow(S.Show)
     const s1 = new Set<string>([])
-    assert.deepStrictEqual(S.show(s1), `new Set([])`)
+    assert.deepStrictEqual(Sh.show(s1), `new Set([])`)
     const s2 = new Set<string>(['a'])
-    assert.deepStrictEqual(S.show(s2), `new Set(["a"])`)
+    assert.deepStrictEqual(Sh.show(s2), `new Set(["a"])`)
     const s3 = new Set<string>(['a', 'b'])
-    assert.deepStrictEqual(S.show(s3), `new Set(["a", "b"])`)
+    assert.deepStrictEqual(Sh.show(s3), `new Set(["a", "b"])`)
   })
 })